lesson56/requirements.txt
pytest>=8.0.0
selenium>=4.20.0
Appium-Python-Client>=4.0.0
